--- src/by/it/rudzko/jd02_01/IUseBasket.java
@@ -1,9 +1,0 @@
-package by.it.rudzko.jd02_01;
-
-/**
- * Created by Leta on 04.10.2016.
- */
-public interface IUseBasket {
-    void takeBasket();
-    void putGoods();
-}
